Instrument GraphQL requests
